test(osapifix): rename typedassert testdata to src/v4/test1.go
