+++ cmd/osapifix/testdata/src/v4/test1.go
@@ -4,7 +4,7 @@
 // this file be licensed under the Apache-2.0 license or a
 // compatible open source license.
 
-package a
+package v4
 
 import (
 	"github.com/stretchr/testify/require"
