+++ cmd/osapifix/typedassert_test.go
@@ -12,10 +12,10 @@ import (
 	"golang.org/x/tools/go/analysis/analysistest"
 )
 
-// TestTypedAssertAnalyzer runs the analyzer against testdata/src/a, whose
+// TestTypedAssertAnalyzer runs the analyzer against testdata/src/v4, whose
 // //want comments mark the lines that must be flagged. Lines without a //want
 // comment must NOT be flagged (false positives fail the test just as missed
 // diagnostics do).
 func TestTypedAssertAnalyzer(t *testing.T) {
-	analysistest.Run(t, analysistest.TestData(), TypedAssertAnalyzer, "a")
+	analysistest.Run(t, analysistest.TestData(), TypedAssertAnalyzer, "v4")
 }
